fix: clear remaining build warnings

+++ src/core/config.cc
@@ -181,8 +181,8 @@ InputPhrase Preferences::sanitizeInputPhrase( QString const & inputPhrase ) cons
 
   if( limitInputPhraseLength && inputPhrase.size() > inputPhraseLengthLimit )
   {
-    gdDebug( "Ignoring an input phrase %d symbols long. The configured maximum input phrase length is %d symbols.",
-             inputPhrase.size(), inputPhraseLengthLimit );
+    gdDebug( "Ignoring an input phrase %lld symbols long. The configured maximum input phrase length is %d symbols.",
+         static_cast<long long>( inputPhrase.size() ), inputPhraseLengthLimit );
     return result;
   }
